Fixed premissions issues. Also removed forbidden buttons

+++ Gosh/Controllers/CategoriesController.cs
@@ -22,10 +22,6 @@ namespace Gosh.Controllers
         {
             return View(db.Categories.ToList());
         }
-        public bool IsAdmin()
-        {
-            return Session["Username"] != null && Session["Username"].ToString() == "ADMIN";
-        }
         // GET: User
 
 
@@ -64,7 +60,7 @@ namespace Gosh.Controllers
         // GET: Categories/Create
         public ActionResult Create()
         {
-            if (!IsAdmin())
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
             {
                 return RedirectToAction("Forbidden", "User");
             }
@@ -78,7 +74,7 @@ namespace Gosh.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ID,CategoryName,ImagePath,RepresetingArea,WeatherHref")] Category category, HttpPostedFileBase file)
         {
-            if (!IsAdmin())
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
             {
                 return RedirectToAction("Forbidden", "User");
             }
@@ -104,9 +100,9 @@ namespace Gosh.Controllers
         // GET: Categories/Edit/5
         public ActionResult Edit(long? id)
         {
-            if (!IsAdmin())
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
             {
-                return RedirectToAction("Forbidden","User");
+                return RedirectToAction("Forbidden", "User");
             }
             if (id == null)
             {
@@ -127,6 +123,10 @@ namespace Gosh.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ID,CategoryName,ImagePath,RepresetingArea,WeatherHref")] Category category)
         {
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
+            {
+                return RedirectToAction("Forbidden", "User");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(category).State = EntityState.Modified;
@@ -139,7 +139,7 @@ namespace Gosh.Controllers
         // GET: Categories/Delete/5
         public ActionResult Delete(long? id)
         {
-            if (!IsAdmin())
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
             {
                 return RedirectToAction("Forbidden", "User");
             }
@@ -160,6 +160,10 @@ namespace Gosh.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(long id)
         {
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
+            {
+                return RedirectToAction("Forbidden", "User");
+            }
             Category category = db.Categories.Find(id);
             db.Categories.Remove(category);
             db.SaveChanges();
