+++ Gosh/Controllers/RecipesController.cs
@@ -92,6 +92,10 @@ namespace Gosh.Controllers
         // GET: Recipes/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
+            {
+                return RedirectToAction("Forbidden", "User");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -111,6 +115,11 @@ namespace Gosh.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "RecipeId,DateCreated,Header,Summary,Content,HomeImageUrl,CategoryId")] Recipe recipe)
         {
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
+            {
+                return RedirectToAction("Forbidden", "User");
+            }
+
             if (ModelState.IsValid)
             {
                 db.Entry(recipe).State = EntityState.Modified;
@@ -132,7 +141,14 @@ namespace Gosh.Controllers
             {
                 return HttpNotFound();
             }
+
+            if(Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
+            {
+                return RedirectToAction("Forbidden", "User");
+                
+            }
             return View(recipe);
+
         }
 
         // POST: Recipes/Delete/5
@@ -140,6 +156,10 @@ namespace Gosh.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Session["Username"] == null || Session["Username"].ToString() != "ADMIN")
+            {
+                return RedirectToAction("Forbidden", "User");
+            }
             Recipe recipe = db.Recipes.Find(id);
             db.Recipes.Remove(recipe);
             db.SaveChanges();
